added game class and output
dataframe output
