--- test.py
@@ -1,5 +1,0 @@
-import random
-
-a = [11.0222]
-print(float(a))
-print(a)
